% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_data_by_month.R
\name{get_account_data_by_month}
\alias{get_account_data_by_month}
\title{Get data for specifics accounts in specified months}
\usage{
get_account_data_by_month(.data = NULL, month = 12, match_required = TRUE)
}
\arguments{
\item{.data}{vector indicating the account, Null for all accounts}

\item{month}{vector indicating the months of reference.}

\item{match_required}{logic indicates if the account names must match}
}
\value{
tibble with account data.
}
\description{
Get data for specifics accounts in specified months
}
\examples{
get_account_data_by_month(.data= "5. RESULTADO PRIMÁRIO GOVERNO CENTRAL - ACIMA DA LINHA (3 - 4)", month = 12)
}
