% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expense_data.R
\name{get_expense_data}
\alias{get_expense_data}
\title{Get RTN expense data by year}
\usage{
get_expense_data(year = NULL)
}
\arguments{
\item{year}{vector indicating the year or reference. NULL for all years}
}
\value{
tibble with all.
}
\description{
Get RTN expense data by year
}
\examples{
get_expense_data(year = 2021)
}
