% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_full_account_name.R
\name{get_full_account_name}
\alias{get_full_account_name}
\title{Get the full account name related to a given part of an account name}
\usage{
get_full_account_name(account = NULL)
}
\arguments{
\item{account}{vector indicating part of the account name, Null for all accounts}
}
\value{
tibble with account data.
}
\description{
Get the full account name related to a given part of an account name
}
\examples{
get_full_account_name(account= "resultado")
}
