% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_full_data.R
\name{get_full_data}
\alias{get_full_data}
\title{Get all the available data about RTN}
\usage{
get_full_data()
}
\value{
tibble with all.
}
\description{
Get all the available data about RTN
}
\examples{
get_full_data()
}
