% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_revenue_data.R
\name{get_revenue_data}
\alias{get_revenue_data}
\title{Get RTN revenue data by year}
\usage{
get_revenue_data(year = NULL)
}
\arguments{
\item{year}{vector indicating the year or reference. NULL for all years}
}
\value{
tibble with all.
}
\description{
Get RTN revenue data by year
}
\examples{
get_revenue_data(year = 2021)
}
