% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_year_accumulated_account_data.R
\name{get_year_accumulated_account_data}
\alias{get_year_accumulated_account_data}
\title{Get account data accumulated in a year for specifics accounts in specified months}
\usage{
get_year_accumulated_account_data(.data = NULL, match_required = TRUE)
}
\arguments{
\item{.data}{vector indicating the account, Null for all accounts}

\item{match_required}{logic indicates if the account names must match}
}
\value{
tibble with account data.
}
\description{
Get account data accumulated in a year for specifics accounts in specified months
}
\examples{
get_year_accumulated_account_data(.data= "5. RESULTADO PRIMÁRIO GOVERNO CENTRAL - ACIMA DA LINHA (3 - 4)")
}
