% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_12_month_accumulated.R
\name{graph_12_month_accumulated}
\alias{graph_12_month_accumulated}
\title{Generate a graph showing a the evolution of 12 month accumulated values}
\usage{
graph_12_month_accumulated(.data, value_type = "1")
}
\arguments{
\item{.data}{tibble with 12 month accumulated values}

\item{value_type}{character 1- For updated value}
}
\value{
line graph.
}
\description{
Generate a graph showing a the evolution of 12 month accumulated values
}
\examples{

}
