% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_rtn_series.R
\name{graph_rtn_series}
\alias{graph_rtn_series}
\title{Generate a graph showing a the evolution of 12 month accumulated values}
\usage{
graph_rtn_series(.data, value_type = "1", clean_names = TRUE)
}
\arguments{
\item{.data}{tibble with 12 month accumulated values}

\item{value_type}{character 1- For updated value}
}
\value{
line graph.
}
\description{
Generate a graph showing a the evolution of 12 month accumulated values
}
\examples{
get_12_month_accumulated_account_data_by_month(c("3. RECEITA LÃƒÂQUIDA","despesa total","acima da linha"), month = c(1:12), match_required = FALSE) \%>\% graph_rtn_series()
}
