% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rtn_series.R
\name{plot_rtn_series}
\alias{plot_rtn_series}
\title{Generate a line graph for a given accounts data}
\usage{
plot_rtn_series(.data, value_type = "1", clean_names = TRUE)
}
\arguments{
\item{.data}{tibble with 12 month accumulated values}

\item{value_type}{character 1- For updated value}
}
\value{
line graph.
}
\description{
Generate a line graph for a given accounts data
}
\examples{
get_year_accumulated_account_data(c("3. RECEITA LÍQUIDA","despesa total","acima da linha"), match_required = FALSE) \%>\% plot_rtn_series()
}
