% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seasonality.R
\name{plot_seasonality}
\alias{plot_seasonality}
\title{Generates a seasonality graphic for a given account data}
\usage{
plot_seasonality(.data, value_type = "1", clean_names = TRUE, polar = TRUE)
}
\arguments{
\item{.data}{tibble with 12 month accumulated values}

\item{value_type}{character 1- For updated value}

\item{clean_names}{logical uses polar coordinates}
}
\value{
line graph.
}
\description{
Generates a seasonality graphic for a given account data
}
\examples{
get_account_data_by_month(c("física"), month = c(1:12), match_required = FALSE) \%>\% plot_seasonality()
}
