\name{tni.graph}
\alias{tni.graph}

\title{
Compute a graph from TNI objects (Deprecated).
}

\description{
Extract results from a TNI object and compute a graph (Deprecated).
}

\usage{
tni.graph(x, ...)
}

\arguments{
\item{x}{Deprecated arg.}
\item{...}{Additional deprecated args.}
}

\value{
---
}

\examples{
# deprecated function

}

\keyword{methods}
