\name{upgradeTNA}
\alias{upgradeTNA}

\title{Upgrade objects of class TNI.}

\description{
This function provides compatibility checks for a TNI class object.
}

\usage{
upgradeTNA(object)
}

\arguments{
\item{object}{
this argument is an object of class \code{\link[RTN:TNI-class]{TNI-class}}.
}
}

\value{
An updated TNA-class object.
}

\author{Mauro Castro}

\examples{

### Objects of class TNA generated by RTN (version <= 1.15.2) can be upgraded
### to the latest version by calling upgradeTNA().
}

\keyword{upgrade}

