\name{upgradeTNI}
\alias{upgradeTNI}

\title{Upgrade objects of class TNI.}

\description{
This function provides compatibility checks for a TNI class object.
}

\usage{
upgradeTNI(object)
}

\arguments{
\item{object}{
this argument is an object of class \code{\link[RTN:TNI-class]{TNI-class}}.
}
}

\value{
An updated TNI-class object.
}

\author{Mauro Castro}

\examples{

### Objects of class TNI generated by RTN (version <= 1.15.2) can be upgraded
### to the latest version by calling upgradeTNI().
}

\keyword{upgrade}

