% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seasonality.R
\name{plot_seasonality}
\alias{plot_seasonality}
\title{Generate a line graph for a given accounts data}
\usage{
plot_seasonality(.data, value_type = "1", clean_names = TRUE, polar = TRUE)
}
\arguments{
\item{.data}{tibble with 12 month accumulated values}

\item{value_type}{character 1- For updated value}
}
\value{
line graph.
}
\description{
Generate a line graph for a given accounts data
}
\examples{
graph_rtn_series(c("3. RECEITA LÃƒÂQUIDA","despesa total","acima da linha"), month = c(1:12), match_required = FALSE) \%>\% graph_rtn_series()
}
